setup _layouts page to load font and same UI as app.js
